send asset changes

+++ neo/Implementations/Wallets/peewee/UserWallet.py
@@ -28,6 +28,7 @@ from neo.Implementations.Wallets.peewee.Models import Account, Address, Coin, Co
 
 from autologging import logged
 import json
+from playhouse.migrate import *
 
 @logged
 class UserWallet(Wallet):
@@ -51,6 +52,14 @@ class UserWallet(Wallet):
         except Exception as e:
             print("Couldnt build database %s " % e)
             self.__log.debug("couldnt build database %s " % e)
+
+    def Migrate(self):
+        db = PWDatabase.ContextDB()
+        migrator = SqliteMigrator(db)
+
+        migrate(
+            migrator.drop_not_null('Contract','Account_id')
+        )
 
     def DB(self):
         return PWDatabase.Context()
@@ -319,10 +328,10 @@ class UserWallet(Wallet):
             for ct in self._contracts.values():
                 if ct.PublicKeyHash == k.PublicKeyHash:
                     signature_contract = ct
+            if signature_contract:
+                addr = signature_contract.Address
 
-            addr = signature_contract.Address
-
-            jsn.append( {'Address': addr, 'Public Key': pub.decode('utf-8')})
+                jsn.append( {'Address': addr, 'Public Key': pub.decode('utf-8')})
 
         return jsn
 
@@ -339,8 +348,30 @@ class UserWallet(Wallet):
                 print("Couldnt delete coin %s %s " % (e, coin))
                 self.__log.debug("could not delete coin %s %s " % (coin, e))
 
-        address = Address.get(ScriptHash = bytes(script_hash.ToArray()))
-        address.delete_instance()
+
+        print("script hash %s %s " % ( script_hash, type(script_hash)))
+
+        todelete = bytes(script_hash.ToArray())
+        print("to delete: %s " % todelete)
+        for c in Contract.select():
+
+            address = c.Address
+            if address.ScriptHash == todelete:
+
+                c.delete_instance()
+                address.delete_instance()
+                print("deleting contract!!!")
+#            print("SCRIPT HASH %s " % c.ScriptHash)
+#            if c.ScriptHash == todelete:
+#                c.delete_instance()
+#        contract = Contract.get(ScriptHash=bytes(script_hash.ToArray()))
+#        print("contract to delete %s " % contract)
+#        contract.delete_instance()
+        try:
+            address = Address.get(ScriptHash = todelete)
+            address.delete_instance()
+        except Exception as e:
+            pass
 
         return True,coins_toremove
